Move helper function to util package
