+++ pkg/util/pki/pki_common_test.go
@@ -175,38 +175,3 @@ func TestControllerUserForCluster(t *testing.T) {
 		t.Errorf("Expected %+v\nGot %+v", expected, user)
 	}
 }
-
-func TestTruncatedCommonName(t *testing.T) {
-	testCases := []struct {
-		testName           string
-		commonName         string
-		commonNameLenLimit int
-		expected           string
-	}{
-		{
-			testName:           "Common name with length exceeded limitation",
-			commonName:         "kafka-test-controller-exceeded-length.test-namespace.mgt.cluster.local",
-			commonNameLenLimit: MaxCertManagerCNLen,
-			expected:           "kafka-test-controller-exceeded-length.test-namespace.mgt.cluster",
-		},
-		{
-			testName:           "Common name with length within limitation",
-			commonName:         "kafka-test-controller.test-namespace.mgt.cluster.local",
-			commonNameLenLimit: MaxCertManagerCNLen,
-			expected:           "kafka-test-controller.test-namespace.mgt.cluster.local",
-		},
-	}
-
-	t.Parallel()
-
-	for _, test := range testCases {
-		test := test
-
-		t.Run(test.testName, func(t *testing.T) {
-			get := TruncatedCommonName(test.commonName, test.commonNameLenLimit)
-			if get != test.expected {
-				t.Errorf("Expected CN name after truncation:%s, got:%s", test.expected, get)
-			}
-		})
-	}
-}
